2025exercises/07-smolagents/requirements.txt
smolagents[toolkit]
requests
python-dotenv
gradio
